feat. type 모음
